pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docval"
version = "0.1.1"
description = "Validate and refactor Markdown documentation against source code using heuristics + LLM"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    {name = "Tom Sapletta", email = "tom@sapletta.com"},
]
keywords = ["documentation", "validation", "refactoring", "llm", "markdown", "static-analysis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Documentation",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
    "gitpython>=3.1",
    "goal>=2.1.0",
    "costs>=0.1.20",
    "pfix>=0.1.60",
]

[project.optional-dependencies]
llm = ["litellm>=1.0"]
all = ["docval[llm]"]
dev = ["pytest>=7.0", "pytest-cov>=4.0", "ruff>=0.1",
    "goal>=2.1.0",
    "costs>=0.1.20",
    "pfix>=0.1.60",
]

[project.scripts]
docval = "docval.cli:main"

[project.urls]
Homepage = "https://github.com/wronai/docval"
Repository = "https://github.com/wronai/docval"

[tool.setuptools.packages.find]
where = ["src"]
include = ["docval*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.pfix]
# Self-healing Python configuration
model = "openrouter/qwen/qwen3-coder-next"
auto_apply = true
auto_install_deps = true
auto_restart = false
max_retries = 3
create_backups = false
git_auto_commit = false

[tool.pfix.runtime_todo]
enabled = true
todo_file = "TODO.md"
min_severity = "low"
deduplicate = true

[tool.costs]
# AI Cost tracking configuration
badge = true
update_readme = true
readme_path = "README.md"
default_model = "openrouter/qwen/qwen3-coder-next"
analysis_mode = "byok"
full_history = true
max_commits = 500

# Cost thresholds for badge colors (USD)
badge_color_thresholds = { low = 1.0, medium = 5.0, high = 10.0, critical = 50.0 }
